pyproject.toml
[project]
name = "payment-processor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic==1.18.4",
    "asyncpg==0.31.0",
    "fastapi[standard]==0.136.0",
    "faststream[rabbit]==0.6.7",
    "loguru>=0.7.3",
    "orjson>=3.11.8",
    "pydantic-settings>=2.13.1",
    "sqlalchemy==2.0.49",
    "starlette-exporter>=0.23.0",
]

[dependency-groups]
dev = [
    "mypy==1.20.1",
    "pre-commit==4.5.1",
    "pytest-asyncio==1.3.0",
    "ruff==0.15.11",
]

[tool.mypy]
python_version = 3.13
plugins = "pydantic.mypy"
ignore_missing_imports = true

# https://docs.pydantic.dev/latest/integrations/mypy/#enabling-the-plugin
follow_imports = "silent"
# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = false

[[tool.mypy.overrides]]
# Дока - https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml-file
module = "dateutil.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiofiles"
ignore_missing_imports = true

[tool.ruff]
# https://beta.ruff.rs/docs/rules/
lint.select = [
    "E", # PyCodeStyle errors
    "W", # PyCodeStyle warnings
    "I", # isort
    "N", # pep8-naming
    "F", # Pyflakes
    "UP", # pyupgrade
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "B", # flake8-bugbear
    "ASYNC", # flake8-asynс
    "G", # flake8-logging-forma
    "DTZ", # flake8-datetimez раскоментить после рефакторинг
    "D", # - докстринги - самый душный линтер
    "RUF", # Ruff-specific rules
    "ANN", # - neccecary type hints (undocument foo)
]
lint.ignore = [
    "ANN002", #  ANN002 Missing type annotations for `*a`
    "ANN003", #  ANN003 Missing type annotations for `**kw`
    "ANN401", # ANN401 Dynamically typed expressions (typing.Any) are disallowed in `record`

    "D100", # D100 Missing docstring in public module
    "D101", # D101 Missing docstring in public class
    "D104", #  D104 Missing docstring in public package
    "D105", # D105 Missing docstring in magic method
    "D106", # D106 Missing docstring in public nested class
    "D107", # Missing docstring in `__init__`
    "D202", # D202cNo blank lines allowed after function docstring

    "N805", # N805 First argument of a method should be named `self`
    "B008", # B008 Do not perform function call `Depends` in argument defaults
    "RUF001", # String contains ambiguous
    "RUF002", # Docstring contains ambiguous
    "RUF003", # Comment contains ambiguous
    "RUF012", # Mutable class attributes should be ___annotated_types with `typing.ClassVar`
    "W291", # Trailing whitespace
]
lint.extend-select = [
    "B006", # mutable-argument-default
    "PIE794", # duplicate-class-field-definition
]

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Показать все исправленное флагом --fix
show-fixes = true


# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
    "env",
    ".js",
    "migration",
    "migrations"
]

# Same as Black.
line-length = 120

# Assume Python 3.13.
target-version = "py313"

[tool.ruff.lint.isort]
force-wrap-aliases = true
force-single-line = true
combine-as-imports = false
lines-between-types = 1

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[tool.ruff.lint.per-file-ignores]
"migrations/*.py" = ["D", "E402", "E501", "ANN"]
"tests/*.py" = ["D", "DTZ", "E501", "ANN"]
"tests/conftest.py" = ["E402"]
"notebooks/*.ipynb" = ["T201", "T203", "F405", "D", "ANN", "F403"]
"tests/libraries/schemas/source_systems/*.py" = ["N815"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false

# Настройки покрытия
[tool.coverage.run]
branch = true
source = ["app"]
omit = [
    "*/__init__.py",
    "tests/*",
    "*/tests/*",
    "app/main.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ['tests']
pythonpath = '.'
